scripts/requirements.txt
mistralai>=1.0.0
pillow>=10.0.0